djangorest/requirements.txt
django
gunicorn
django-heroku==0.3.1
djangorestframework